Fix O -> A make function and update tests.
Running tests revealed a “bug” in the make function, that we were
assuming the existence of a “completed_at” property in the Asana Task
Node even though this property isn’t guaranteed to exist.  In practice
all Asana Tasks currently come with a “completed_at” property, but meh.
Can’t hurt to make the code match the tests, and it’s just a check on
inputs anyway.

There’s also a real fix, to make sure the Asana Project ID is used
appropriately.