BUG FIX: accept make headline return value
Somehow it took me this long to discover that when Emacs makes a new
headline, the return value back to Python is not a string containing
just the new id, but rather a string containing the new id *in double
quotes*.  So we’ve gotta “eval” that real quick.

Updated tests to reflect reality, and fixed the code so all tests pass.