Use Org’s “todo type” instead of “todo keyword” throughout.
The type is more universally reliable.  It’s always either todo or done,
whereas the keyword can be just about anything the user defines.

There’s some easy-to-confuse name overloading going on in here.
Technically, in elisp the :todo-type of a headline is either
the *symbol* todo or the *symbol* done, and in either case you shouldn’t
set it directly.  So (1) we’re passing it to and fro as a string and (2)
when setting it in elisp we’re actually setting the todo keyword.