Asana: share tag info via a dedicated dictionary-like object.
The old approach to sharing tag info across all Asana nodes relied on a
shared class variable (dictionary).  This dirty hack didn’t work well
with testing, so now we’re doing it the right way with a shared
dictionary-like object.

The object, TagNameLookup(), also knows how to create new tags on the
server.  Kind of an RAII approach.  Seems to work for now.