Test side effects of conforming two trees.
Hey, at least we’ve got some tests now!  Eventually I’ll muster the
energy to test other side effects, like … does it throw out the right
commands in the right order?  Ugh … testing procedural side-effect code
is the worst.

Recall that it would be infinitely better if the edit_script produced a
list of commands that could be executed later a/k/a if the edit_script
were more functional and less procedural.  I tried to go down that path
before but got stuck somewhere (I think it was hard to calculate the
correct indices for some operations?) so I gave up.  Maybe if I write
enough tests I’ll have the confidence to just go do it.